add `prefer-import-meta-properties`, and revert `prefer-module`
