+++ rules/prefer-module.js
@@ -1,23 +1,12 @@
-import {findVariable} from '@eslint-community/eslint-utils';
 import isShadowed from './utils/is-shadowed.js';
 import assertToken from './utils/assert-token.js';
 import {getCallExpressionTokens} from './utils/index.js';
-import {
-	isStaticRequire,
-	isReferenceIdentifier,
-	isFunction,
-	isMemberExpression,
-	isCallExpression,
-	isNewExpression,
-	isMethodCall,
-} from './ast/index.js';
+import {isStaticRequire, isReferenceIdentifier, isFunction} from './ast/index.js';
 import {removeParentheses, replaceReferenceIdentifier, removeSpacesAfter} from './fix/index.js';
 
 const ERROR_USE_STRICT_DIRECTIVE = 'error/use-strict-directive';
 const ERROR_GLOBAL_RETURN = 'error/global-return';
 const ERROR_IDENTIFIER = 'error/identifier';
-const ERROR_CALCULATE_DIRNAME = 'error/calculate-dirname';
-const ERROR_CALCULATE_FILENAME = 'error/calculate-filename';
 const SUGGESTION_USE_STRICT_DIRECTIVE = 'suggestion/use-strict-directive';
 const SUGGESTION_IMPORT_META_DIRNAME = 'suggestion/import-meta-dirname';
 const SUGGESTION_IMPORT_META_URL_TO_DIRNAME = 'suggestion/import-meta-url-to-dirname';
@@ -29,8 +18,6 @@ const messages = {
 	[ERROR_USE_STRICT_DIRECTIVE]: 'Do not use "use strict" directive.',
 	[ERROR_GLOBAL_RETURN]: '"return" should be used inside a function.',
 	[ERROR_IDENTIFIER]: 'Do not use "{{name}}".',
-	[ERROR_CALCULATE_DIRNAME]: 'Do not construct dirname.',
-	[ERROR_CALCULATE_FILENAME]: 'Do not construct filename using `fileURLToPath()`.',
 	[SUGGESTION_USE_STRICT_DIRECTIVE]: 'Remove "use strict" directive.',
 	[SUGGESTION_IMPORT_META_DIRNAME]: 'Replace `__dirname` with `import.meta.dirname`.',
 	[SUGGESTION_IMPORT_META_URL_TO_DIRNAME]: 'Replace `__dirname` with `…(import.meta.url)`.',
@@ -211,155 +198,6 @@ const isTopLevelReturnStatement = node => {
 	return true;
 };
 
-const isParentLiteral = node => {
-	if (node?.type !== 'Literal') {
-		return false;
-	}
-
-	return node.value === '.' || node.value === './';
-};
-
-const isImportMeta = node =>
-	node.type === 'MetaProperty'
-	&& node.meta.name === 'import'
-	&& node.property.name === 'meta';
-
-function isNodeBuiltinModuleFunctionCall(node, {modules, functionName, sourceCode}) {
-	if (!isCallExpression(node, {optional: false, argumentsLength: 1})) {
-		return false;
-	}
-
-	const visited = new Set();
-
-	return checkExpression(node.callee, 'property');
-
-	/** @param {import('estree').Expression} node */
-	function checkExpression(node, checkKind) {
-		if (node.type === 'MemberExpression') {
-			if (!(
-				checkKind === 'property'
-				&& isMemberExpression(node, {property: functionName, computed: false, optional: false})
-			)) {
-				return false;
-			}
-
-			return checkExpression(node.object, 'module');
-		}
-
-		if (node.type === 'CallExpression') {
-			if (checkKind !== 'module') {
-				return false;
-			}
-
-			// `process.getBuiltinModule('x')`
-			return (
-				isMethodCall(node, {
-					object: 'process',
-					method: 'getBuiltinModule',
-					argumentsLength: 1,
-					optionalMember: false,
-					optionalCall: false,
-				})
-				&& isModuleLiteral(node.arguments[0])
-			);
-		}
-
-		if (node.type !== 'Identifier') {
-			return false;
-		}
-
-		if (visited.has(node)) {
-			return false;
-		}
-
-		visited.add(node);
-
-		const variable = findVariable(sourceCode.getScope(node), node);
-		if (!variable || variable.defs.length !== 1) {
-			return;
-		}
-
-		return checkDefinition(variable.defs[0], checkKind);
-	}
-
-	/** @param {import('eslint').Scope.Definition} define */
-	function checkDefinition(define, checkKind) {
-		if (define.type === 'ImportBinding') {
-			if (!isModuleLiteral(define.parent.source)) {
-				return false;
-			}
-
-			const specifier = define.node;
-			return checkKind === 'module'
-				? (specifier?.type === 'ImportDefaultSpecifier' || specifier?.type === 'ImportNamespaceSpecifier')
-				: (specifier?.type === 'ImportSpecifier' && specifier.imported.name === functionName);
-		}
-
-		return define.type === 'Variable' && checkPattern(define.name, checkKind);
-	}
-
-	/** @param {import('estree').Identifier | import('estree').ObjectPattern} node */
-	function checkPattern(node, checkKind) {
-		/** @type {{parent?: import('estree').Node}} */
-		const {parent} = node;
-		if (parent.type === 'VariableDeclarator') {
-			if (
-				!parent.init
-				|| parent.id !== node
-				|| parent.parent.type !== 'VariableDeclaration'
-				|| parent.parent.kind !== 'const'
-			) {
-				return false;
-			}
-
-			return checkExpression(parent.init, checkKind);
-		}
-
-		if (parent.type === 'Property') {
-			if (!(
-				checkKind === 'property'
-				&& parent.value === node
-				&& !parent.computed
-				&& parent.key.type === 'Identifier'
-				&& parent.key.name === functionName
-			)) {
-				return false;
-			}
-
-			// Check for ObjectPattern
-			return checkPattern(parent.parent, 'module');
-		}
-
-		return false;
-	}
-
-	function isModuleLiteral(node) {
-		return node?.type === 'Literal' && modules.has(node.value);
-	}
-}
-
-/**
-@returns {node is import('estree').SimpleCallExpression}
-*/
-function isUrlFileURLToPathCall(node, sourceCode) {
-	return isNodeBuiltinModuleFunctionCall(node, {
-		modules: new Set(['url', 'node:url']),
-		functionName: 'fileURLToPath',
-		sourceCode,
-	});
-}
-
-/**
-@returns {node is import('estree').SimpleCallExpression}
-*/
-function isPathDirnameCall(node, sourceCode) {
-	return isNodeBuiltinModuleFunctionCall(node, {
-		modules: new Set(['path', 'node:path']),
-		functionName: 'dirname',
-		sourceCode,
-	});
-}
-
 function fixDefaultExport(node, sourceCode) {
 	return function * (fixer) {
 		yield fixer.replaceText(node, 'export default ');
@@ -519,141 +357,6 @@ function create(context) {
 		}
 
 		return problem;
-	});
-
-	context.on('MetaProperty', function * (node) {
-		if (!isImportMeta(node)) {
-			return;
-		}
-
-		/** @type {import('estree').Node} */
-		const memberExpression = node.parent;
-		if (!isMemberExpression(memberExpression, {
-			properties: ['url', 'filename'],
-			computed: false,
-			optional: false,
-		})) {
-			return;
-		}
-
-		const propertyName = memberExpression.property.name;
-		if (propertyName === 'url') {
-			// `url.fileURLToPath(import.meta.url)`
-			if (
-				isUrlFileURLToPathCall(memberExpression.parent, sourceCode)
-				&& memberExpression.parent.arguments[0] === memberExpression
-			) {
-				yield * iterateProblemsFromFilename(memberExpression.parent, {
-					reportFilenameNode: true,
-				});
-				return;
-			}
-
-			// `new URL(import.meta.url)`
-			// `new URL('.', import.meta.url)`
-			// `new URL('./', import.meta.url)`
-			if (isNewExpression(memberExpression.parent, {name: 'URL', minimumArguments: 1, maximumArguments: 2})) {
-				const newUrl = memberExpression.parent;
-				const urlParent = newUrl.parent;
-
-				// `new URL(import.meta.url)`
-				if (newUrl.arguments.length === 1 && newUrl.arguments[0] === memberExpression // `url.fileURLToPath(new URL(import.meta.url))`
-
-					&& isUrlFileURLToPathCall(urlParent, sourceCode)
-					&& urlParent.arguments[0] === newUrl
-				) {
-					yield * iterateProblemsFromFilename(urlParent, {
-						reportFilenameNode: true,
-					});
-					return;
-				}
-
-				// `url.fileURLToPath(new URL(".", import.meta.url))`
-				// `url.fileURLToPath(new URL("./", import.meta.url))`
-				if (
-					newUrl.arguments.length === 2
-					&& isParentLiteral(newUrl.arguments[0])
-					&& newUrl.arguments[1] === memberExpression
-					&& isUrlFileURLToPathCall(urlParent, sourceCode)
-					&& urlParent.arguments[0] === newUrl
-				) {
-					yield getProblem(urlParent, 'dirname');
-				}
-			}
-
-			return;
-		}
-
-		if (propertyName === 'filename') {
-			yield * iterateProblemsFromFilename(memberExpression);
-		}
-
-		/**
-		Iterates over reports where a given filename expression node
-		would be used to convert it to a dirname.
-		@param { import('estree').Expression} node
-		*/
-		function * iterateProblemsFromFilename(node, {reportFilenameNode = false} = {}) {
-			/** @type {{parent: import('estree').Node}} */
-			const {parent} = node;
-
-			// `path.dirname(filename)`
-			if (
-				isPathDirnameCall(parent, sourceCode)
-				&& parent.arguments[0] === node
-			) {
-				yield getProblem(parent, 'dirname');
-				return;
-			}
-
-			if (reportFilenameNode) {
-				yield getProblem(node, 'filename');
-			}
-
-			if (
-				parent.type !== 'VariableDeclarator'
-				|| parent.init !== node
-				|| parent.id.type !== 'Identifier'
-				|| parent.parent.type !== 'VariableDeclaration'
-				|| parent.parent.kind !== 'const'
-			) {
-				return;
-			}
-
-			/** @type {import('eslint').Scope.Variable|null} */
-			const variable = findVariable(sourceCode.getScope(parent.id), parent.id);
-			if (!variable) {
-				return;
-			}
-
-			for (const reference of variable.references) {
-				if (!reference.isReadOnly()) {
-					continue;
-				}
-
-				/** @type {{parent: import('estree').Node}} */
-				const {parent} = reference.identifier;
-				if (
-					isPathDirnameCall(parent, sourceCode)
-					&& parent.arguments[0] === reference.identifier
-				) {
-					// Report `path.dirname(identifier)`
-					yield getProblem(parent, 'dirname');
-				}
-			}
-		}
-
-		/**
-		@param { import('estree').Node} node
-		@param {'dirname' | 'filename'} name
-		*/
-		function getProblem(node, name) {
-			return {
-				node,
-				messageId: name === 'dirname' ? ERROR_CALCULATE_DIRNAME : ERROR_CALCULATE_FILENAME,
-				fix: fixer => fixer.replaceText(node, `import.meta.${name}`),
-			};
-		}
 	});
 }
