+++ scripts/create-rule.js
@@ -1,12 +1,13 @@
 #!/usr/bin/env node
 import fs from 'node:fs';
 import path from 'node:path';
+import {fileURLToPath} from 'node:url';
 import enquirer from 'enquirer';
 import {template} from 'lodash-es';
 import openEditor from 'open-editor';
 import spawn from 'nano-spawn';
 
-const {dirname} = import.meta;
+const dirname = path.dirname(fileURLToPath(import.meta.url));
 const ROOT = path.join(dirname, '..');
 
 function checkFiles(ruleId) {
